requirements.txt
# Django Framework Requirements
# Core Django and REST API
Django>=5.1.7
djangorestframework>=3.15.0
django-cors-headers>=4.4.0
django-filter>=23.5.0

# Database drivers
mysqlclient>=2.2.0
PyMySQL>=1.1.0

# Machine Learning - Core
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8 framework
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
opencv-python-headless>=4.8.0

# Visualization & Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Web Framework Extensions
gunicorn>=21.2.0
whitenoise>=6.6.0

# API Documentation
drf-yasg>=1.21.7
coreapi>=2.3.3

# File Handling & Storage
django-storages>=1.14.2
boto3>=1.34.0  # For AWS S3 storage if needed

# Background Tasks (Optional)
celery>=5.3.0
redis>=5.0.0

# Monitoring & Logging
django-extensions>=3.2.3
django-debug-toolbar>=4.2.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.57.0

# Weather API Integration
requests>=2.31.0
python-decouple>=3.8

# Data Processing & Analysis
joblib>=1.3.0
psutil>=5.9.0  # For system monitoring

# Development & Testing (Optional)
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
coverage>=7.3.0

# Production Dependencies
uwsgi>=2.0.23  # Alternative to gunicorn
supervisor>=4.2.5  # Process management

# Additional ML Libraries (Optional but recommended)
xgboost>=2.0.0  # For ensemble forecasting
lightgbm>=4.1.0  # Alternative gradient boosting
optuna>=3.4.0  # Hyperparameter optimization

# Time Series Forecasting (for pest forecasting)
statsmodels>=0.14.0
pmdarima>=2.0.4

# Geospatial Support (for location-based features)
geopy>=2.4.0
folium>=0.15.0  # For mapping visualizations

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.66.0  # Progress bars for long operations
colorlog>=6.8.0  # Colored logging

# File Format Support
openpyxl>=3.1.0  # Excel file support
xlrd>=2.0.1
xlwt>=1.3.0

# Configuration Management
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Performance Optimization
django-cache-machine>=1.2.0
django-cachalot>=2.6.1

# Monitoring and Health Checks
django-health-check>=3.17.0

# API Rate Limiting
django-ratelimit>=4.1.0

# Content Management
django-ckeditor>=6.7.0  # Rich text editor for admin

# Export/Import Support
django-import-export>=3.3.0

# Mobile App Support
django-push-notifications>=3.0.2  # For mobile alerts

# Documentation Generation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Code Quality (Development)
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
pre-commit>=3.5.0

# Type Checking (Development)
mypy>=1.7.0
django-stubs>=4.2.7

# Security
django-security>=0.17.0
bandit>=1.7.5  # Security linting

# Backup and Migration Tools
django-dbbackup>=4.0.2

# Email Support
django-ses>=3.5.0  # AWS SES email backend

# Internationalization
django-modeltranslation>=0.18.11

# GraphQL API (Optional)
graphene-django>=3.2.0

# WebSocket Support (for real-time updates)
channels>=4.0.0
channels-redis>=4.1.0

# Environment-specific requirements
# Production optimizations
django-silk>=5.0.4  # Performance profiling
django-prometheus>=2.3.1  # Metrics collection

# Note: Some packages may have conflicts or require specific versions
# Install core requirements first, then add optional packages as needed
# For production, pin exact versions for stability
